refactor(errors): reorganize error handling by moving to core module and updating imports
